Add test route to backend
