+++ client/package.json
@@ -48,5 +48,6 @@
     "tslint-config-prettier": "^1.18.0",
     "tslint-plugin-prettier": "^2.0.1",
     "tslint-react": "^4.1.0"
-  }
+  },
+  "proxy": "http://localhost:3000/"
 }
