+++ client/tslint.json
@@ -13,7 +13,8 @@
       "member-ordering": false,
       "object-literal-sort-keys": false,
       "ordered-imports": false,
-      "ban-types": true
+      "ban-types": true,
+      "no-console": false
     },
     "linterOptions": {
       "exclude": [
