+++ server/package.json
@@ -15,8 +15,10 @@
     "@server": "dist/Server"
   },
   "dependencies": {
+    "@types/cors": "^2.8.6",
     "@types/oracledb": "^4.0.5",
     "cookie-parser": "^1.4.4",
+    "cors": "^2.8.5",
     "cross-env": "^6.0.3",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
