+++ server/src/routes/index.ts
@@ -1,11 +1,11 @@
 import { Router } from 'express';
-import UserRouter from './Users';
+import IncidentRouter from './Incidents';
 
 // Init router and path
 const router = Router();
 
 // Add sub-routes
-router.use('/users', UserRouter);
+router.use('/incidents', IncidentRouter);
 
 // Export the base-router
 export default router;
